% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\name{san_fran_baker_beach}
\alias{san_fran_baker_beach}
\title{san_fran_baker_beach}
\usage{
san_fran_baker_beach(location = c("Baker Beach West", "Baker Beach East",
  "Baker Beach at Lobos Creek"), status = FALSE)
}
