% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\name{san_fran_china_beach}
\alias{san_fran_china_beach}
\title{san_fran_china_beach}
\usage{
san_fran_china_beach(status = FALSE)
}
\arguments{
\item{status}{. Show beach status?}
}
\description{
Downloads and tidies the San Francisco Public Utilities Commission Beach Water Quality Monitoring Program data. Datasets can be downloaded per beach, or the raw data can be downloaded.
}
\examples{
\dontrun{
china <- san_fran_china_beach()

# See if the beach is open:
san_fran_china_beach(status = TRUE)
}
}
