% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\name{san_fran_crissy_field}
\alias{san_fran_crissy_field}
\title{san_fran_crissy_field}
\usage{
san_fran_crissy_field(location = c("West", "East"), status = FALSE)
}
\arguments{
\item{location}{. Choose location.}

\item{status}{. Show beach status?}
}
\description{
Downloads and tidies the San Francisco Public Utilities Commission Beach Water Quality Monitoring Program data. Datasets can be downloaded per beach, or the raw data can be downloaded.
}
