% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\name{san_fran_mission_creek}
\alias{san_fran_mission_creek}
\title{san_fran_mission_creek}
\usage{
san_fran_mission_creek(status = FALSE)
}
\arguments{
\item{status}{. Show beach status?}
}
\description{
Downloads and tidies the San Francisco Public Utilities Commission Beach Water Quality Monitoring Program data. Datasets can be downloaded per beach, or the raw data can be downloaded.
}
