% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\name{san_fran_Ocean_Beach}
\alias{san_fran_Ocean_Beach}
\title{san_fran_Ocean_Beach}
\usage{
san_fran_Ocean_Beach(location = c("Sloat Boulevard", "Lincoln Way",
  "Balboa Street"), status = FALSE)
}
\arguments{
\item{location}{. Choose location.}

\item{status}{. Show beach status?}
}
\description{
Downloads and tidies the San Francisco Public Utilities Commission Beach Water Quality Monitoring Program data. Datasets can be downloaded per beach, or the raw data can be downloaded.
}
