% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaches.R
\name{san_fran_raw_data}
\alias{san_fran_raw_data}
\title{Download San Francisco Beach Water Quality Data}
\usage{
san_fran_raw_data()
}
\description{
Downloads and tidies the San Francisco Public Utilities Commission Beach Water Quality Monitoring Program data. Datasets can be downloaded per beach, or the raw data can be downloaded.
}
